requirements.txt
# Core Dependencies
Flask==3.1.0
Flask-CORS==5.0.0
Flask-SQLAlchemy==3.1.1
tensorflow-cpu==2.20.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn==1.6.1

# Database
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
alembic==1.15.1

# Monitoring & Logging
prometheus-client==0.21.1
python-json-logger==3.2.1
APScheduler==3.11.0

# API & Security
gunicorn==23.0.0
python-dotenv==1.0.1
PyJWT==2.10.1
werkzeug==3.1.5

# Data Processing
scipy==1.15.1
joblib==1.4.2

# Testing & Performance
pytest==8.3.4
locust==2.33.0
requests==2.32.5

# Visualization
plotly==6.0.0
matplotlib==3.10.0

# Utilities
python-dateutil==2.9.0
pytz==2025.1